remove pricer hook
