+++ src/hooks/data/vaults/fallbacks/pricer/usePricer.ts
@@ -1,16 +1,16 @@
 import { useQuery } from "@tanstack/react-query"
 import { Address } from "wagmi"
 
-import { getAladdinApiPrice } from "@/hooks/data/vaults/fallbacks/pricer/useAladdinApiPricer"
-import { getApyVisionApiPrice } from "@/hooks/data/vaults/fallbacks/pricer/useApyVisionApiPricer"
-import { getCurveFactoryCryptoPrice } from "@/hooks/data/vaults/fallbacks/pricer/useCurveFactoryCryptoPricer"
-import { getCurveFactoryPrice } from "@/hooks/data/vaults/fallbacks/pricer/useCurveFactoryPricer"
-import { getCurveLpTokenPrice } from "@/hooks/data/vaults/fallbacks/pricer/useCurveLpTokenPricer"
-import { getCurveMainPrice } from "@/hooks/data/vaults/fallbacks/pricer/useCurveMainPricer"
-import { getGlpPrice } from "@/hooks/data/vaults/fallbacks/pricer/useGlpPricer"
-import { getLlamaApiPrice } from "@/hooks/data/vaults/fallbacks/pricer/useLlamaApiPricer"
-import { getLlamaArbiApiPrice } from "@/hooks/data/vaults/fallbacks/pricer/useLlamaArbiApiPricer"
-import { getLlamaEthPrice } from "@/hooks/data/vaults/fallbacks/pricer/useLlamaEthPricer"
+import { getAladdinApiPrice } from "@/lib/api/pricer/getAladdinApiPrice"
+import { getApyVisionApiPrice } from "@/lib/api/pricer/getApyVisionApiPrice"
+import { getCurveFactoryCryptoPrice } from "@/lib/api/pricer/getCurveFactoryCryptoPrice"
+import { getCurveFactoryPrice } from "@/lib/api/pricer/getCurveFactoryPrice"
+import { getCurveLpTokenPrice } from "@/lib/api/pricer/getCurveLpTokenPrice"
+import { getCurveMainPrice } from "@/lib/api/pricer/getCurveMainPrice"
+import { getGlpPrice } from "@/lib/api/pricer/getGlpPrice"
+import { getLlamaApiPrice } from "@/lib/api/pricer/getLlamaApiPrice"
+import { getLlamaArbiApiPrice } from "@/lib/api/pricer/getLlamaArbiApiPrice"
+import { getLlamaEthPrice } from "@/lib/api/pricer/getLlamaEthPrice"
 
 import { PRIMARYASSET_PRICER } from "@/constant/mapping"
