+++ src/lib/api/pricer/getCurveFactoryCryptoPrice.ts
@@ -1,5 +1,4 @@
 import axios from "axios"
-import { Address, useQuery } from "wagmi"
 
 import { CURVE_FACTORY_CRYPTO_URL } from "@/constant/env"
 
@@ -12,18 +11,4 @@ export async function getCurveFactoryCryptoPrice(token: string) {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const relevantCoin = coins.filter((x: any) => x?.["address"] === token)
   return relevantCoin?.[0]?.["usdPrice"]
-}
-
-export default function useCurveFactoryCryptoPricer({
-  primaryAsset,
-  enabled,
-}: {
-  primaryAsset: Address | undefined
-  enabled: boolean
-}) {
-  return useQuery(["curveFactoryCryptoPricer", primaryAsset ?? "0x"], {
-    queryFn: () => getCurveFactoryCryptoPrice(primaryAsset ?? "0x"),
-    retry: false,
-    enabled: enabled,
-  })
 }
